cjdns: python2 -> python3
The `contrib` folder was removed from the packaging. It contains a lot of files
that are irrelevant to the actual package (i.e. Fedora spec files, selinux hardening
or systemd services). There are a few scripts in there too but in any case they are
misplaced and would be better off in a separate package (as e.g. the Fedora packaging
does)