python3Packages.pipBuildHook: prepare for building in-place
Pip for a while copied the source files to a temporary directory and
built there. It won't do that anymore in the future. Enable the feature
to test for issues. In the future we need to revert this commit.

DEPRECATION: A future pip version will change local packages to be built
in-place without first copying to a temporary directory. We recommend
you use --use-feature=in-tree-build to test your packages with this new
behavior before it becomes the default.
pip 21.3 will remove support for this functionality. You can find
discussion regarding this at https://github.com/pypa/pip/issues/7555.