beets: 1.5.0 -> 1.6.0
Refresh patches, remove keyfinder patch as upstream seems to use the
correct keyfinder-cli path, since:
https://github.com/beetbox/beets/pull/3467/files

Remove the doing nothing `imagemagick` in `buildInputs` and add a patch
that makes `beets` use `magick` from Nix' path. The patch can be
disabled along with the reference to imagemagick if it's set to `null`.

Formatting: use 1 line per input.

Remove the unevaluated externalTestArgs.beets - it was moved to
`all-packages.nix`.