vscode-extensions.ms-vscode.cpptools: fix wrapping of OpenDebugAD7
`OpenDebugAD7` is no longer packaged as a .NET binary but a native
Linux one. I have introduced `autoPatchelfHook` to be just done with
the wrapping.

`liblttng-ust.so` had to be replaced as the binaries were depending on
a different version. This is done similarly as is e.g. with the
`powershell` derivation.