GnuCash: refactor in support of enabling GSettings
GnuCash uses Gnome's GSettings to store user preferences. This was not
working before. wrapGAppsHook is introduced to solve this problem, but
it must be tweaked to not wrap GnuCash's cli utilities for fetching
financial quotes. Those are now deliberately wrapped with their Perl
library dependencies.

The existing code is refactored in an attempt to clarify which
dependencies are needed and why.