nixos/tests/mtp: use QEMU v6.0.0
A change in QEMU v6.1.0 has somehow caused QEMU to behave differently
enough to cause this test to fail. This commit forces the test to be ran
with QEMU 6.0.0 from Nixpkgs at revision
e1fc1a80a071c90ab65fb6eafae5520579163783, which is the commit prior to
the QEMU 6.1.0 version bump.

Co-authored-by: Julio Sueiras <juliosueiras@gmail.com>