helix: reenable haskell syntax highlighting support
Haskell syntax highlighting is disabled in `helix-editor` by default.
This is because of an issue compiling tree-sitter-haskell on M1 macs,
as its default compiler doesn't support c++14.
As we're using a nix-derived toolchain, we shouldn't be subject to
this problem.

See [this comment](https://github.com/helix-editor/helix/issues/1384#issuecomment-1001240194)
for more info.