matrix-appservice-slack: Slim down dependencies
Due to the setup of the node env and npm packages in general,
the resulting nix derivations contain a lot of unwanted dependencies.
In the the case of matrix-appservice-slack, this includes a fully
functional gcc which is required to build native bindings but is
unnecessary at runtime, as well as a full nodejs toolchain,
including tooling like npm and the like.

Switching to nodejs-slim and patching over node paths, in addition with
stripping debug symbols from native bindings, mitigates this a lot,
resulting in a massively reduced closure.