elvish: properly set buildinfo via `ldflags`
As of [this commit][1] (v0.16.0), the module path of Elvish changed from
`github.com/elves/elvish` to `src.elv.sh`. This is also reflected in the
updated [packaging instructions][2].

This commit updates the `ldflags` in the derivation to use the new
module path so that `buildinfo.Reproducible` is correctly set to `true`.

[1]: https://github.com/elves/elvish/commit/196eea21d4f185d6ac203e7f9a4fa9a9c4d680f4
[2]: https://github.com/elves/elvish/blob/master/PACKAGING.md