fetchFromGitHub: name source with content intristic to the value
Just giving the derivation a name of "source" by default causes
a lot of pain when someone may update the rev, but forget to
fail to compute the new sha256 or invalidate the sha256. A stale
sha256 will still be "valid" in the eyes of nix, as only the way
in which the sources are fetched has changed, but the previous
.drv and sha256 relationship is still preserved.

This change forces the FODs to be recomputed as the related .drv
has changed.

This will not require all of the FOD sha256s to be recomputed, as
the contents will not have changed. However, it will be a mass
rebuild as it will create a new DAG for packages which didn't
already pass a "name" value to fetchFromGitHub.