vigra: python2 -> python3
This derivation already doesn't build vigra's python bindings; the only
use of python is the `bin/vigra-config` script that dependent packages
can use to get compiler flags. I confirmed that this script works
correctly on python3. Without building the python bindings, the numpy
dependency is not used, so was removed.

A `diff` between the old and new output directories shows that the only
changes are in `bin/vigra-config`: the `#!` path to python and the
absolute paths to the output directory that would change on any rebuild.