Fix loading of libcrypto when using salt-ssh
When using salt-ssh, Salt would copy itself to the target machine. The libcrypto loading patch included in Nix hardcodes the location, but does not provide any fallback for the target machine, making salt-ssh unusable from a nix environment, which is ironically, one of the coolest use-cases of nix: a self contained git repo with a Saltfile and a shell.nix with dependencies.