mkYarnPackage: Change default arguments
The packageJSON default was overwritten by literally every package using it, since
the default produces IFD which is forbidden in nixpkgs. `./package.json` would be
a good default, but that doesn't work because of the nature of relative paths.

The yarnLock argument was overwritten in about 70% of the packages (but to different
values respectively), so the default value was removed.

mkYarnWorkspace was changed the same way but it doesn't matter because nobody uses it.
(What is even a workspace, in that context?)