WIP: Start decoupling X and graphics support with `videoDrivers`
Now that Wayland is less unviable, we should begin to refactor our
options so you don't need to use X11-named options just to do graphics
without X11.

This will be a long process, and I am not really sure where to start,
but `videoDrivers` seems like a decent option.