Idris2: Use derivation provided by idris2 project
Since we created the nix derivation in nixpkgs, the idris2 project
has incorporated our derivation and improved on it. By now, the
derivation managed by the project has many more features than ours.
Furthermore, it would be unwise for us not to reuse the derivation
that is provided by the project maintainers themselves.

Improvements on the derivation should be done upstream in the
project, whenever possible, instead of overwriting them here.