tagtime: init at 2018-09-02
The original perl implementation of tagtime.

I packaged it to the point of being able to use it on my system, in a
very barebones state.
The configuration is … well it just assumes you have a perl file
somewhere and it evals it at startup. Since that configuration needs
access to the distpath of tagtime, we just patch it a bit for
now (which also means no user configuration is possible atm).

The main changes to the config are:

- instead of defaulting to `vim`, we use `$EDITOR`
- The log file is placed in `$HOME/.local/share/tagtime/pings.log`
  (which corresponds to `$XDG_DATA_HOME` on normal unix systems), but
  could be patched to use the real logic instead.