fetchCargoTarball: make it work with sandbox
The call to `tar -z` in fetchCargoTarball causes `tar` to execve
`/bin/sh -c gzip`. But with sandboxing enabled, this results in a ENOENT
no such file or directory error on `/bin/sh`. We can work around this by
not letting `tar` call `gzip`, but calling it ourselves without the
intermediate `/bin/sh` step, and then it does work with the sandbox.

Fixes #153774.