mautrix-facebook: Remove max version requirements.
It appears that these are bumped to match manual testing however they do not work well in nixpkgs as every time a dependency is updated this package breaks. To resolve this issue just strip the max version requirements and assume that it works.

Ideally we would do a semver-type comparison but this is likely not an issue because a new major version would need a new attribute name anyways.