gnome.gnome-shell: switch to non-full asciidoc
The full variant does not seem to be actually needed,
allowing us to prune Java from the build closure.

In fact, the AsciiDoc/a2x tool is not actually used
at all when building from tarball because it contains
a pre-generated copy of the manual page.

Let’s remove the pregenerated file so that building
from tarball is more consistent with the VCS build
(e.g. when testing new versions or patches).