tracker: Unconditionally add checkInputs to nativeBuildInputs
`gi` is required by the main `meson.build` and checked at the start
of the build process, regardless of whether tests are enabled or not.
As a result, builds are broken when `doCheck = false`.

Also removing tappy which is no longer needed: https://gitlab.gnome.org/GNOME/tracker/-/merge_requests/385