Idris2: Split builtin packages into outputs
We now offer idris2Packages similar to idris. The set idris2Packages
now offers all builtin packages like base, contrib and, network
separately. We also offer a with-packages, which allows to easily
merge multiple idris2 packages with the idris2 binary.

This is the first step to bring a proper package infrastructure to
idris2. We do not, yet, offer a nice way to compile idris2 packages to
be used with with-packages, yet.